+++ src/translations/en/twigpack.php
@@ -2,7 +2,8 @@
 /**
  * Twigpack plugin for Craft CMS 3.x
  *
- * Twigpack is the conduit between Twig and webpack, with manifest.json & webpack-dev-server HMR support
+ * Twigpack is the conduit between Twig and webpack, with manifest.json &
+ * webpack-dev-server HMR support
  *
  * @link      https://nystudio107.com/
  * @copyright Copyright (c) 2018 nystudio107
@@ -15,6 +16,7 @@
  */
 return [
     'Twigpack Manifest Cache' => 'Twigpack Manifest Cache',
-    '{name} plugin loaded'    => '{name} plugin loaded',
-    'Manifest file not found at: {manifestPath}' => 'Manifest file not found at: {manifestPath}'
+    '{name} plugin loaded' => '{name} plugin loaded',
+    'Manifest file not found at: {manifestPath}' => 'Manifest file not found at: {manifestPath}',
+    'Module does not exist in the manifest: {moduleName}' => 'Module does not exist in the manifest: {moduleName}',
 ];