+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Twigpack Changelog
 
+## 1.0.4 - 2018-09-28
+### Added
+* Added `this.onload=null;` to async CSS link tag
+* Added `craft.twigpack.includeCssRelPreloadPolyfill()`
+
+### Changed
+* Better error reporting if modules don't exist in the manifest
+
+## 1.0.3 - 2018-09-24
+### Changed
+* Allow the `manifestPath` to be a file system path or a URI
+
 ## 1.0.2 - 2018-09-23
 ### Added
 * Added `getModuleUri()` function